Renamed TableControllerV3 class as TableController.
