+++ src/components/account/AccountTableController.ts
@@ -21,10 +21,10 @@
 import {AccountInfo, AccountsResponse} from "@/schemas/HederaSchemas";
 import {ComputedRef} from "vue";
 import axios, {AxiosResponse} from "axios";
-import {KeyOperator, SortOrder, TableControllerV3} from "@/utils/table/TableControllerV3";
+import {KeyOperator, SortOrder, TableController} from "@/utils/table/TableController";
 import {Router} from "vue-router";
 
-export class AccountTableController extends TableControllerV3<AccountInfo, string> {
+export class AccountTableController extends TableController<AccountInfo, string> {
 
     //
     // Public
