+++ src/components/block/BlockTableController.ts
@@ -21,10 +21,10 @@
 import {Block, BlocksResponse} from "@/schemas/HederaSchemas";
 import {ComputedRef} from "vue";
 import axios, {AxiosResponse} from "axios";
-import {KeyOperator, SortOrder, TableControllerV3} from "@/utils/table/TableControllerV3";
+import {KeyOperator, SortOrder, TableController} from "@/utils/table/TableController";
 import {Router} from "vue-router";
 
-export class BlockTableController extends TableControllerV3<Block, number> {
+export class BlockTableController extends TableController<Block, number> {
 
     //
     // Public
