+++ src/components/token/TokenBalanceTableController.ts
@@ -21,10 +21,10 @@
 import {TokenBalancesResponse, TokenDistribution} from "@/schemas/HederaSchemas";
 import {ComputedRef, Ref} from "vue";
 import axios, {AxiosResponse} from "axios";
-import {KeyOperator, SortOrder, TableControllerV3} from "@/utils/table/TableControllerV3";
+import {KeyOperator, SortOrder, TableController} from "@/utils/table/TableController";
 import {Router} from "vue-router";
 
-export class TokenBalanceTableController extends TableControllerV3<TokenDistribution, string> {
+export class TokenBalanceTableController extends TableController<TokenDistribution, string> {
 
     public readonly tokenId: Ref<string | null>
