+++ src/components/transaction/TransactionTableController.ts
@@ -18,14 +18,14 @@
  *
  */
 
-import {KeyOperator, SortOrder, TableControllerV3} from "@/utils/table/TableControllerV3";
+import {KeyOperator, SortOrder, TableController} from "@/utils/table/TableController";
 import {Transaction, TransactionResponse} from "@/schemas/HederaSchemas";
 import {ComputedRef, ref, Ref} from "vue";
 import axios, {AxiosResponse} from "axios";
 import {Router} from "vue-router";
 
 
-export class TransactionTableController extends TableControllerV3<Transaction, string> {
+export class TransactionTableController extends TableController<Transaction, string> {
 
     private readonly accountId: Ref<string | null>
     private readonly accountIdMandatory: boolean
